+++ MapboxSearch/place-autocomplete/src/test/java/com/mapbox/search/autocomplete/PlaceAutocompleteSuggestionTest.kt
@@ -1,6 +1,5 @@
 package com.mapbox.search.autocomplete
 
-import com.mapbox.geojson.Point
 import com.mapbox.search.autocomplete.test.utils.TypeObjectCreator
 import com.mapbox.search.common.tests.CommonSdkTypeObjectCreators
 import com.mapbox.search.common.tests.ReflectionObjectsFactory
@@ -8,7 +7,6 @@ import com.mapbox.search.common.tests.ToStringVerifier
 import com.mapbox.search.common.tests.withPrefabTestBoundingBox
 import com.mapbox.search.common.tests.withPrefabTestPoint
 import nl.jqno.equalsverifier.EqualsVerifier
-import org.junit.jupiter.api.Assertions.assertSame
 import org.junit.jupiter.api.Test
 
 internal class PlaceAutocompleteSuggestionTest {
@@ -28,35 +26,11 @@ internal class PlaceAutocompleteSuggestionTest {
             objectsFactory = ReflectionObjectsFactory(
                 extraCreators = CommonSdkTypeObjectCreators.ALL_CREATORS +
                         TypeObjectCreator.OPEN_HOURS_CREATOR +
-                        TypeObjectCreator.PLACE_TYPE_CREATOR
+                        TypeObjectCreator.PLACE_TYPE_CREATOR +
+                        TypeObjectCreator.SUGGESTION_UNDERLYING_CREATOR
             ),
-            ignoredProperties = listOf("result"),
+            ignoredProperties = listOf("underlying"),
             includeAllProperties = false
         ).verify()
-    }
-
-    @Test
-    fun `result() function is correct`() {
-        val result = PlaceAutocompleteResult(
-            name = "Test result",
-            coordinate = Point.fromLngLat(10.0, 15.0),
-            routablePoints = null,
-            makiIcon = "test",
-            distanceMeters = 123.0,
-            address = null,
-            type = PlaceAutocompleteType.AdministrativeUnit.Place,
-            categories = null,
-            phone = "+123456789",
-            website = "https://test.com",
-            reviewCount = 123,
-            averageRating = 5.0,
-            openHours = null,
-            primaryPhotos = listOf(),
-            otherPhotos = listOf(),
-        )
-
-        val suggestion = PlaceAutocompleteSuggestion(result)
-
-        assertSame(result, suggestion.result())
     }
 }